Modified PacketHandler (now ModNetwork) to be more consistent with typical network standards
